dev: - splitted options dataclass from main code to separate file
